requirements.txt
absl-py==2.1.0
albucore==0.0.9
albumentations==1.4.8
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
efficientnet==1.0.0
efficientnet_pytorch==0.6.3
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
huggingface-hub==0.23.3
idna==3.7
image==1.5.33
image-classifiers==1.0.0
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
munch==4.0.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
patchify==0.2.3
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pretrainedmodels==0.7.4
prompt-toolkit==3.0.42
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
# pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
segmentation-models==1.0.1
segmentation-models-pytorch==0.2.0
setuptools==69.5.1
six==1.16.0
split-folders==0.5.1
sqlparse==0.5.0
stack-data==0.6.2
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==0.4.12
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
zipp==3.17.0
